Separate SIKE and SIDH
