+++ CMakeLists.txt
@@ -67,8 +67,11 @@ endif()
 if(OQS_ENABLE_KEM_FRODOKEM)
     set(PUBLIC_HEADERS ${PUBLIC_HEADERS} ${PROJECT_SOURCE_DIR}/src/kem/frodokem/kem_frodokem.h)
 endif()
-if(OQS_ENABLE_KEM_SIKE OR OQS_ENABLE_KEM_SIDH)
+if(OQS_ENABLE_KEM_SIKE)
     set(PUBLIC_HEADERS ${PUBLIC_HEADERS} ${PROJECT_SOURCE_DIR}/src/kem/sike/kem_sike.h)
+endif()
+if(OQS_ENABLE_KEM_SIDH)
+    set(PUBLIC_HEADERS ${PUBLIC_HEADERS} ${PROJECT_SOURCE_DIR}/src/kem/sike/kem_sidh.h)
 endif()
 if(OQS_ENABLE_SIG_PICNIC)
     set(PUBLIC_HEADERS ${PUBLIC_HEADERS} ${PROJECT_SOURCE_DIR}/src/sig/picnic/sig_picnic.h)
