+++ src/kem/sike/P751/P751.c
@@ -135,6 +135,9 @@ static const unsigned int strat_Bob[MAX_Bob - 1] = {
 #define crypto_kem_keypair OQS_KEM_sike_p751_keypair
 #define crypto_kem_enc OQS_KEM_sike_p751_encaps
 #define crypto_kem_dec OQS_KEM_sike_p751_decaps
+#define OQS_KEM_sidh_keypair OQS_KEM_sidh_p751_keypair
+#define OQS_KEM_sidh_encaps OQS_KEM_sidh_p751_encaps
+#define OQS_KEM_sidh_decaps OQS_KEM_sidh_p751_decaps
 #define random_mod_order_A oqs_kem_sidh_p751_random_mod_order_A
 #define random_mod_order_B oqs_kem_sidh_p751_random_mod_order_B
 #define EphemeralKeyGeneration_A oqs_kem_sidh_p751_EphemeralKeyGeneration_A
@@ -156,4 +159,4 @@ static const unsigned int strat_Bob[MAX_Bob - 1] = {
 #include "../fpx.c"
 #include "../ec_isogeny.c"
 #include "../sidh.c"
-#include "../sike.c"
+#include "../kem.c"
